requirements.txt
gradio==4.16.0
crewai==0.11.0
langchain==0.1.10
langchain-openai==0.0.5
langchain-community==0.0.25
openai==1.12.0
python-dotenv==1.0.0
tavily-python==0.3.0